Grib1Collection : implement tinvFilter  improve Grib FC docs

+++ grib/src/main/java/ucar/nc2/grib/grib2/Grib2CollectionBuilder.java
@@ -467,7 +467,7 @@ public class Grib2CollectionBuilder {
           gr.setFile(fileno); // each record tracks which file it belongs to
           int gdsHash = gr.getGDSsection().getGDS().hashCode();  // use GDS hash code to group records
           if (gdsConvert != null && gdsConvert.get(gdsHash) != null) // allow external config to muck with gdsHash. Why? because of error in encoding
-            gdsHash = (Integer) gdsConvert.get(gdsHash);             // and we need exact hash matching
+            gdsHash = gdsConvert.get(gdsHash);             // and we need exact hash matching
 
           Group g = gdsMap.get(gdsHash);
           if (g == null) {
@@ -519,12 +519,6 @@ public class Grib2CollectionBuilder {
         prob = (int) (1000 * pds.getProbabilityUpperLimit());
       }
       return intvFilter.filterOut(id, haveLength, prob);
-      //Integer needLength = intvFilter.getLengthById(id);
-
-      //if (needLength != null && needLength != haveLength) {
-        //f.format(" FILTER INTV [%d != %d] %s%n", haveLength, needLength, gr);
-        //return true;
-      //}
     }
     return false;
   }
